requirements.txt
asgiref==3.7.2
click==8.1.7
colorama==0.4.6
Django==5.0.1
django-debug-toolbar==4.2.0
dlib==19.24.2
face-recognition==1.3.0
face_recognition_models==0.3.0
numpy==1.26.3
opencv-python==4.9.0.80
pillow==10.2.0
pybase64==1.3.2
python-dotenv==1.0.0
setuptools==69.0.3
sqlparse==0.4.4
tzdata==2023.4
